make carousel reusable
